+++ src/TestHtmlPdfCliPlus/HtmlPdfShrPlus/HtmlPdfResultTests.cs
@@ -4,8 +4,6 @@
 // https://github.com/FRACerqueira/HtmlPdfPlus
 // ***************************************************************************************
 
-using System.IO.Compression;
-using System.Text;
 using System.Text.Json;
 using HtmlPdfPlus;
 using HtmlPdfPlus.Shared.Core;
@@ -65,48 +63,5 @@ namespace TestHtmlPdfPlus.HtmlPdfShrPlus
             Assert.False(roundtripped.Error.Retryable);
         }
 
-        [Fact]
-        public void DecompressOutputData_ShouldDecompressByteArray()
-        {
-            // Arrange
-            var isSuccess = true;
-            var bufferDrained = false;
-            var elapsedTime = TimeSpan.FromSeconds(1);
-            var originalData = Encoding.UTF8.GetBytes("Test Data");
-            var compressedData = Compress(originalData);
-            var result = new HtmlPdfResult<byte[]>(isSuccess, bufferDrained, elapsedTime, compressedData);
-
-            // Act
-            var decompressedResult = result.DecompressOutputData();
-
-            // Assert
-            Assert.Equal(originalData, decompressedResult.OutputData);
-        }
-
-        [Fact]
-        public void DecompressOutputData_ShouldThrowInvalidOperationException_WhenOutputDataIsNotByteArray()
-        {
-            // Arrange
-            var isSuccess = true;
-            var bufferDrained = false;
-            var elapsedTime = TimeSpan.FromSeconds(1);
-            var outputData = "Test Data";
-            var result = new HtmlPdfResult<string>(isSuccess, bufferDrained, elapsedTime, outputData);
-
-            // Act & Assert
-            Assert.Throws<InvalidOperationException>(() => result.DecompressOutputData());
-        }
-
-        private byte[] Compress(byte[] data)
-        {
-            using (var memoryStream = new System.IO.MemoryStream())
-            {
-                using (var gzipStream = new GZipStream(memoryStream, CompressionMode.Compress))
-                {
-                    gzipStream.Write(data, 0, data.Length);
-                }
-                return memoryStream.ToArray();
-            }
-        }
     }
 }